Re pace-neutrons/Herbert#28 Asynchronous receive gets last message of list of similar.

+++ _test/test_mpi/MessagesCppMPI_tester.m
@@ -9,12 +9,12 @@ classdef MessagesCppMPI_tester < MessagesCppMPI
         function obj = MessagesCppMPI_tester()
             obj = obj@MessagesCppMPI('test_mode');
             % make tester look like 10 workers
-            obj.numLabs_ = 10;
+            obj.numLabs_ = uint64(10);
         end
         
         function [labNum,nLabs]=get_lab_index(obj)
             [labNum,nLabs] = obj.read_cpp_comm_pull_info();
-            obj.numLabs_ = 10;
+            obj.numLabs_ = uint64(10);
         end
     end
 end
